modif theme
Former-commit-id: 41b9363e96bccebcd930a0b6a429c46104a4ff26

+++ htdocs/bimplogistique/objects/BL_CommandeFournReception.class.php
@@ -1469,7 +1469,7 @@ class BL_CommandeFournReception extends BimpObject
                 $update = true;
             }
 
-            if ((float) round($this->getInitData('total_ttc'), 5) !== round($total_ttc, 5)) {
+            if ((float) round($this->getInitData('total_ttc'), 5) !== (float) round($total_ttc, 5)) {
                 $this->set('total_ttc', $total_ttc);
                 $update = true;
             }
